city

+++ encoding/datasets/cityscapes.py
@@ -68,7 +68,7 @@ class CitySegmentation(BaseDataset):
         img = Image.open(self.images[index]).convert('RGB')
         if self.mode == 'test':
             if self.transform is not None:
-                img = np.array(img)/255
+                img = np.array(img, dtype=np.double)/255
                 img = self.transform(img)
             return img, os.path.basename(self.images[index])
         #mask = self.masks[index]
@@ -83,7 +83,7 @@ class CitySegmentation(BaseDataset):
             mask = self._mask_transform(mask)
         # general resize, normalize and toTensor
         if self.transform is not None:
-            img = np.array(img)/255
+            img = np.array(img, , dtype=np.double)/255
             img = self.transform(img)
         if self.target_transform is not None:
             mask = self.target_transform(mask)
